pyproject.toml
[project]
name = "file-conversion-service"
version = "0.1.0"
description = "A service to convert files to PDF."
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.12"

# Production dependencies
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "boto3==1.34.37",
    "python-dotenv==1.0.0",
    "pydantic==2.4.2",
    "httpx==0.25.1",
    "Pillow==10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "moto[s3,dynamodb]==4.2.14",
    "pytest-cov==4.1.0",
    "python-multipart==0.0.9",
]

[tool.uv]
env = ".venv"

[tool.uv.scripts]
start = "uvicorn main:app --reload"
test = "pytest"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
